Move models from social api entry to models dedicated package
